+++ docs/Changelog.md
@@ -39,6 +39,7 @@ based on merged pull requests. Search GitHub issues and pull requests for smalle
 - Add limits for flex access/egress walking [#4655](https://github.com/opentripplanner/OpenTripPlanner/pull/4655)
 - Make Prometheus endpoint take Accept header into account [#4674](https://github.com/opentripplanner/OpenTripPlanner/pull/4674)
 - Add entrances to parking lots that are unconnected in OSM [#4666](https://github.com/opentripplanner/OpenTripPlanner/pull/4666)
+- Include crossing=traffic_signals and use it for walking/cycling penalty on crossings [#4574](https://github.com/opentripplanner/OpenTripPlanner/pull/4574)
 [](AUTOMATIC_CHANGELOG_PLACEHOLDER_DO_NOT_REMOVE)